packages/ml/requirements.txt
# Core ML libraries
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0

# Deep Learning
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.7.0
pillow>=9.5.0
pymupdf>=1.22.0  # For PDF processing

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# Image processing
scikit-image>=0.20.0
albumentations>=1.3.0

# Vector search
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# API integration
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Testing
pytest>=7.3.0